namwoo_app/requirements.txt
# requirements.txt
# This file lists the Python dependencies for the NamFulgor Flask application.

# --- Core Web Framework ---
Flask>=2.3,<3.0
gunicorn>=21.0.0,<22.0.0

# --- Environment & Configuration ---
python-dotenv>=1.0.0

# --- Database & ORM ---
Flask-SQLAlchemy>=3.0,<3.2
Flask-Migrate>=4.0,<5.0
SQLAlchemy>=2.0,<2.1
psycopg2-binary>=2.9.0,<3.0.0
pgvector>=0.2.0,<0.3.0

# --- AI & HTTP Clients ---
# Used for OpenAI, Azure, and the Google Gemini provider
openai>=1.3.0,<2.0.0
# Used by many services for making API calls
requests>=2.30.0,<3.0.0

# --- Redis Client ---
# Required for Assistant providers (OpenAI & Azure) for locking
redis>=5.0,<6.0